requirements.txt
pillow
numpy
matplotlib
qiskit
qiskit-aer
streamlit